move: Notification 도메인 내 패키지 세분화, member-role 패키지 세분화
