+++ src/main/java/com/coop/domain/comment/service/CommentService.java
@@ -10,7 +10,7 @@ import com.coop.global.common.enums.ErrorCode;
 import com.coop.global.exception.error.AccessDeniedException;
 import com.coop.global.exception.error.InvalidRequestException;
 import com.coop.global.exception.error.NotFoundException;
-import com.coop.global.notification.values.NotificationTarget;
+import com.coop.domain.notification.enums.NotificationTarget;
 import com.coop.global.notification.values.TriggerNotification;
 import com.coop.presentation.comment.dto.request.CommentRequest;
 import com.coop.presentation.comment.dto.response.CommentPageResponse;
