+++ src/main/java/com/coop/domain/notification/enums/NotificationTarget.java
@@ -1,4 +1,4 @@
-package com.coop.global.notification.values;
+package com.coop.domain.notification.enums;
 
 public enum NotificationTarget {
     COMMENT
