+++ src/main/java/com/coop/global/notification/values/NotificationEvent.java
@@ -1,5 +1,6 @@
 package com.coop.global.notification.values;
 
+import com.coop.domain.notification.enums.NotificationTarget;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
