+++ src/main/java/com/coop/global/notification/values/TriggerNotification.java
@@ -1,5 +1,7 @@
 package com.coop.global.notification.values;
 
+import com.coop.domain.notification.enums.NotificationTarget;
+
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
