+++ src/main/java/com/coop/global/security/JwtUtil.java
@@ -1,6 +1,6 @@
 package com.coop.global.security;
 
-import com.coop.domain.member.entity.Role;
+import com.coop.domain.member.enums.Role;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.JwtBuilder;
 import io.jsonwebtoken.Jwts;
